Pass along HTTP options from Proxy to Model
If you use Model.auth('foo', 'bar'), this sets the :options key
in the params to be sent to the server.

However, these params were not passed along to class methods on
Model themselves. This naive patch checks whether the first arg
of the method is an hash... and merges it with the proxied HTTP
options.

It is going to cause trouble, so that is why a big fat FIXME is
left in the code.