Handle the :from parameter correctly
First, remove the http:// prefix relic from the previous
ActiveResource-base implementation.

Hawk, by design, sets a base URL on the connection itself,
as a resource should be accessible only from a well defined
URL base.

Then, move from calculation from an association into the
base model itself. This way you can now specify from which
path component a resource should be fetched.

If you specify a path starting with a slash, that'll be taken
as-is, concatenated with the URL base. If there is no slash,
then the path will be concatenated to the defined resource
base path.