Fix from() endpoint eating by the Finder cc:@mtylty :spaghetti: :unamused:
When .from() is used, the endpoint from which a resource should be
fetched is passed along in the Proxy chain via the `:options` hash
in the `:endpoint` key.

However, the `:options` hash also holds Typhoeus-specific options.

To overcome this, previously `model_path_from` used to delete the
`:endpoint` key from the `:options` hash. But, when a `Proxy` is
memoized in a model instance, deleting the `:endpoint` would also
delete it from the memoized proxy, thus a second request will be
made to the wrong endpoint.

A proper solution would be not to use the `:options` endpoint to
carry along this information. This is a result of the poor param
transport Hawk is currently implementing.

As this code should be overhauled and properly refactored anyway,
as a short term solution the `:endpoint` option is removed before
passing the `:options` hash to Typhoeus.

And here a :spaghetti: icon is more than due. :unamused: